+++ test/triangulation/constrained.jl
@@ -27,7 +27,7 @@ end
 
 @testset "Testing Shewchuk's PSLG example" begin
     pts, C = second_shewchuk_example_constrained()
-    fails = Base.Threads.Atomic{Int64}(0)
+    fails = Base.Threads.Atomic{Int}(0)
     for i in 1:500
         rng = StableRNG(i^6)
         tri = triangulate(pts; edges=C, rng)
@@ -43,7 +43,7 @@ end
         pts = [(2rand(rng) - 1, rand(rng)) for _ in 1:500]
         x = LinRange(-1, 1, 250)
         a = LinRange(0.0, 1.0, 8)
-        C = Set{NTuple{2,Int64}}()
+        C = Set{NTuple{2,Int}}()
         for i in eachindex(a)
             y = a[i] * x .^ 2
             append!(pts, zip(x, y))
@@ -60,7 +60,7 @@ end
         @show i
         rng = StableRNG(i)
         pts = NTuple{2,Float64}[]
-        C = Set{NTuple{2,Int64}}()
+        C = Set{NTuple{2,Int}}()
         j = 1
         for i in 1:100
             push!(pts, (2i / 101 - 1, 2rand(rng) - 1))
